Refactor exception folder
