+++ src/test/java/seedu/address/logic/LogicManagerTest.java
@@ -20,7 +20,7 @@ import seedu.address.logic.commands.CommandResult;
 import seedu.address.logic.commands.ListCommand;
 import seedu.address.logic.commands.exceptions.CommandException;
 import seedu.address.logic.commands.profile.AddProfileCommand;
-import seedu.address.logic.parser.profile.exceptions.ParseException;
+import seedu.address.logic.parser.exceptions.ParseException;
 import seedu.address.model.Model;
 import seedu.address.model.ModelManager;
 import seedu.address.model.ReadOnlyAddressBook;
