+++ src/test/java/seedu/address/logic/parser/CommandParserTestUtil.java
@@ -6,7 +6,7 @@ import seedu.address.logic.commands.Command;
 import seedu.address.logic.parser.profile.AddProfileCommandParser;
 import seedu.address.logic.parser.profile.DeleteProfileCommandParser;
 import seedu.address.logic.parser.profile.EditProfileCommandParser;
-import seedu.address.logic.parser.profile.exceptions.ParseException;
+import seedu.address.logic.parser.exceptions.ParseException;
 
 /**
  * Contains helper methods for testing command parsers.
